Move the utils from pyanaconda.storage
Move the functions from pyanaconda.storage.utils to the Storage module.
