+++ pyanaconda/modules/storage/disk_selection/selection.py
@@ -25,7 +25,7 @@ from pyanaconda.modules.common.constants.objects import DISK_SELECTION
 from pyanaconda.modules.common.errors.storage import UnavailableStorageError
 from pyanaconda.modules.common.structures.validation import ValidationReport
 from pyanaconda.modules.storage.disk_selection.selection_interface import DiskSelectionInterface
-from pyanaconda.storage.utils import check_disk_selection
+from pyanaconda.modules.storage.disk_selection.utils import check_disk_selection
 
 log = get_module_logger(__name__)
