+++ pyanaconda/modules/storage/partitioning/automatic/automatic_partitioning.py
@@ -26,10 +26,9 @@ from pyanaconda.modules.common.structures.partitioning import PartitioningReques
 from pyanaconda.modules.storage.partitioning.automatic.noninteractive_partitioning import \
     NonInteractivePartitioningTask
 from pyanaconda.modules.storage.partitioning.automatic.utils import get_candidate_disks, \
-    schedule_implicit_partitions, schedule_volumes, schedule_partitions
+    schedule_implicit_partitions, schedule_volumes, schedule_partitions, get_pbkdf_args
 from pyanaconda.modules.storage.platform import platform
 from pyanaconda.modules.storage.partitioning.specification import PartSpec
-from pyanaconda.storage.utils import get_pbkdf_args
 from pyanaconda.core.storage import suggest_swap_size
 
 log = get_module_logger(__name__)
